问题解决

+++ src/main/java/com/dx/facade/texas/resp/DxTableConfigRuleResp.java
@@ -164,7 +164,7 @@ public class DxTableConfigRuleResp implements Serializable {
     private Integer seatsNumberLimit;
 
     @ApiModelProperty("时长（小时）")
-    private Integer hourTimeCount;
+    private BigDecimal hourTimeCount;
 
     @ApiModelProperty("单牌桌累计带入上限（倍数）")
     private Integer bringInLimit;
